cli: deploy/remove command, use the stage associated with the env

+++ packages/restlessness-cli/src/commands/remove/index.ts
@@ -3,7 +3,7 @@ import { spawn } from 'child_process';
 import { promises as fs } from 'fs';
 import path from 'path';
 import chalk from 'chalk';
-import { JsonServices, PathResolver } from '@restlessness/core';
+import { JsonServices, PathResolver, JsonEnvs } from '@restlessness/core';
 
 function spawnAsyncWithInheritStdio(command: string, args: any[]): Promise<void> {
   return new Promise((resolve, reject) => {
@@ -43,6 +43,14 @@ export default async (argv: minimist.ParsedArgs) => {
     deploymentEnv = argv.env;
   }
 
+  const jsonEnv = await JsonEnvs.getEntryById(deploymentEnv);
+  if (!jsonEnv) {
+    throw `Cannot find Environment ${deploymentEnv}`;
+  }
+  if (!jsonEnv.stage) {
+    throw `Cannot deploy! Environment ${jsonEnv.id} does not have an associated stage`;
+  }
+
   const outputPath = path.join(PathResolver.getPrjPath, '.serverless-outputs');
   try {
     await fs.rmdir(outputPath, { recursive: true });
@@ -52,7 +60,7 @@ export default async (argv: minimist.ParsedArgs) => {
     const servicePath = path.relative(process.cwd(), path.join(PathResolver.getServicesJsonPath, `${serviceName}.json`));
 
     console.log(chalk.blue('Restlessness:'), 'Removing service', serviceName);
-    const deployArgs = ['--config', servicePath, 'remove', '--stage', deploymentEnv, '--verbose'];
+    const deployArgs = ['--config', servicePath, 'remove', '--stage', jsonEnv.stage, '--verbose'];
     try {
       await spawnAsyncWithInheritStdio('serverless', deployArgs);
     } catch (e) {
